setup: avoid install gtk-theme everytime

+++ setup/hyprland/config/gtk-theme.sh
@@ -1,3 +1,5 @@
-git clone https://github.com/Fausto-Korpsvart/Catppuccin-GTK-Theme.git /tmp/gtk-catppuccin
-/tmp/gtk-catppuccin/themes/install.sh --tweaks macchiato -t lavender
-rm -rf /tmp/gtk-catppuccin
+if [ ! -d "$HOME/.themes/Catppuccin-Purple-Dark-Macchiato" ]; then
+    git clone https://github.com/Fausto-Korpsvart/Catppuccin-GTK-Theme.git /tmp/gtk-catppuccin
+    /tmp/gtk-catppuccin/themes/install.sh --tweaks macchiato -t lavender
+    rm -rf /tmp/gtk-catppuccin
+fi
